requirements.txt
python-multipart==0.0.20
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
aiohttp==3.11.14
fuzzywuzzy==0.18.0
llama-index==0.12.26
llama-index-cli==0.4.1
llama-index-core==0.12.27
llama-index-llms-azure-openai==0.3.2
llama-index-embeddings-azure-openai==0.3.2
jupyter==1.1.1
ipykernel==6.29.5
pymongo==4.11.3
python-Levenshtein==0.27.1
tqdm==4.67.1
pdf2image==1.17.0
PyMuPDF
python-docx==1.1.2
reportlab==4.3.1
docx2txt==0.8
PyPDF2==3.0.1
llama-parse==0.6.21
nest-asyncio==1.6.0
pdfkit==1.0.0
psycopg2-binary==2.9.10
sshtunnel==0.4.0
wkhtmltopdf==0.2
certifi==2025.4.26
llama-index-vector-stores-postgres==0.5.1
llama-index-vector-stores-chroma==0.4.1
llama-index-retrievers-bm25==0.5.2 
fast_langdetect==0.3.2
markdown2==2.5.3